# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/employees": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据查询参数获取员工列表，支持分页、搜索和筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "获取员工列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段 (例如: employeeId, fullName, createdAt)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"desc\"",
                        "description": "排序顺序 ('asc'或'desc')",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词 (匹配姓名、工号)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "在职状态筛选 ('Active'或'Departed')",
                        "name": "employmentStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应，包含员工列表和分页信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.PagedEmployeesData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证或 Token 无效/过期",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从请求体绑定数据并验证，数据保存到数据库。员工工号由系统自动生成。支持设置可选的入职日期（格式：YYYY-MM-DD）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "新增一个员工",
                "parameters": [
                    {
                        "description": "员工信息。包含必填的姓名，可选的手机号、邮箱、部门和入职日期",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateEmployeePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功的员工对象",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Employee"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误、数据校验失败或入职日期格式无效",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证或 Token 无效/过期",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "手机号或邮箱已存在",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/employees/import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "通过上传 CSV 文件批量导入员工。CSV文件必须包含表头：fullName,phoneNumber,email,department,hireDate。列顺序必须一致。fullName为必填，其他字段可为空。hireDate格式为YYYY-MM-DD。支持GBK和UTF-8编码。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "批量导入员工数据 (CSV)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "包含员工数据的 CSV 文件。表头: fullName,phoneNumber,email,department,hireDate",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导入结果摘要，包含成功和失败的详细信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.BatchImportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求错误，例如文件未提供、文件格式错误、CSV表头不匹配或数据格式错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证或 Token 无效/过期",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/employees/{employeeId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据路径参数员工业务工号获取单个员工的完整信息，包含其作为\"办卡人\"和\"当前使用人\"的号码简要列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "获取指定业务工号的员工详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工业务工号",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应，包含员工详情",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.EmployeeDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的员工工号格式 (保留，以防未来有格式校验)",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证或 Token 无效/过期",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "员工未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/employees/{employeeId}/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据员工业务工号更新员工的部门、入职日期、在职状态或离职日期。所有字段都是可选的，至少需要提供一个字段进行更新。入职日期和离职日期格式为 YYYY-MM-DD。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "更新指定业务工号的员工信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工业务工号",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的员工字段。可包含部门、入职日期、在职状态、离职日期",
                        "name": "employeeUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateEmployeePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的员工对象",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Employee"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误、数据校验失败、日期格式无效或业务逻辑错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未认证或 Token 无效/过期",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "员工未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.BatchImportResponse": {
            "type": "object",
            "properties": {
                "errorCount": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.BatchImportErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                },
                "successCount": {
                    "type": "integer"
                }
            }
        },
        "handlers.CreateEmployeePayload": {
            "type": "object",
            "required": [
                "fullName"
            ],
            "properties": {
                "department": {
                    "type": "string",
                    "maxLength": 255
                },
                "email": {
                    "description": "可选，需要是合法的email格式，最大长度255",
                    "type": "string",
                    "maxLength": 255
                },
                "fullName": {
                    "type": "string",
                    "maxLength": 255
                },
                "hireDate": {
                    "description": "入职日期，可选，格式 YYYY-MM-DD",
                    "type": "string"
                },
                "phoneNumber": {
                    "description": "手机号：可选，但如果提供，必须是11位数字",
                    "type": "string"
                }
            }
        },
...
        "handlers.PagedEmployeesData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Employee"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/handlers.PaginationInfo"
                }
            }
        },
...
        "handlers.PaginationInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "models.Employee": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "department": {
                    "description": "部门",
                    "type": "string"
                },
                "email": {
                    "description": "员工邮箱, 可选, 唯一 (NULLS NOT DISTINCT)",
                    "type": "string"
                },
                "employeeId": {
                    "description": "员工业务工号, 例如 EMP0000001",
                    "type": "string"
                },
                "employmentStatus": {
                    "description": "在职状态 (例如: 'Active', 'Departed')",
                    "type": "string"
                },
                "fullName": {
                    "description": "姓名",
                    "type": "string"
                },
                "hireDate": {
                    "description": "入职日期",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phoneNumber": {
                    "description": "员工手机号码, 11位, 可选, 唯一 (NULLS NOT DISTINCT)",
                    "type": "string"
                },
                "terminationDate": {
                    "description": "离职日期",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "可选的成功消息",
                    "type": "string"
                },
                "status": {
                    "description": "例如 \"success\"",
                    "type": "string"
                }
            }
        }
    }
}